chore: move the package to a subfolder
