+++ package.json
@@ -1,22 +1,11 @@
 {
-  "name": "@wildpeaks/snapshot-dom",
-  "version": "2.0.0-rc1",
-  "description": "Converts a DOM element to a JSON tree",
-  "author": "Cecile Muller",
-  "license": "MIT",
-  "keywords": [
-    "wildpeaks",
-    "assert",
-    "dom"
-  ],
-  "homepage": "https://github.com/wildpeaks/package-snapshot-dom#readme",
-  "repository": "https://github.com/wildpeaks/package-snapshot-dom",
-  "bugs": {
-    "url": "https://github.com/wildpeaks/package-snapshot-dom/issues"
-  },
+  "private": true,
   "eslintConfig": {
     "extends": "@wildpeaks/commonjs"
   },
+  "eslintIgnore": [
+    "packages"
+  ],
   "prettier": {
     "printWidth": 120,
     "tabWidth": 4,
@@ -29,13 +18,7 @@
       "devDependencyUpdate": "chore: update ${dependency}"
     }
   },
-  "main": "src/snapshot.js",
-  "files": [
-    "src",
-    "lib"
-  ],
   "scripts": {
-    "lint": "eslint ./src/**/*.js",
     "test": "mocha test/*.spec.js"
   },
   "devDependencies": {
