+++ packages/snapshot-dom/lib/browser.js
