+++ packages/snapshot-dom/lib/node.js
